Move taxonomic branch into component, use in Suggestions.
